modify sortIncludes and remove CtxDep
