+++ cmd/llcppcfg/llcppgcfg/cfg_test_data/bdw-gc/llcppg.cfg
@@ -4,25 +4,25 @@
 	"libs": "$(pkg-config --libs bdw-gc)",
 	"include": [
 		"gc/gc_inline.h",
-		"gc/gc.h",
-		"gc/gc_version.h",
-		"gc/gc_config_macros.h",
-		"gc/gc_tiny_fl.h",
 		"gc_cpp.h",
-		"gc/gc_cpp.h",
 		"gc/gc_backptr.h",
 		"gc/gc_mark.h",
 		"gc.h",
 		"gc/leak_detector.h",
 		"gc/javaxfc.h",
+		"gc/gc_cpp.h",
 		"gc/gc_disclaim.h",
 		"gc/gc_gcj.h",
 		"gc/gc_typed.h",
+		"gc/gc.h",
 		"gc/ec.h",
 		"gc/cord.h",
-		"gc/cord_pos.h",
 		"gc/gc_pthread_redirects.h",
-		"gc/gc_allocator.h"
+		"gc/gc_tiny_fl.h",
+		"gc/gc_version.h",
+		"gc/gc_config_macros.h",
+		"gc/gc_allocator.h",
+		"gc/cord_pos.h"
 	],
 	"deps": null,
 	"trimPrefixes": [],
